% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_DFO_REST.R
\name{get_DFO_REST}
\alias{get_DFO_REST}
\title{get_DFO_REST}
\usage{
get_DFO_REST(
  host = "https://gisp.dfo-mpo.gc.ca",
  service = "FGP/ADAPT_Canada_Atlantic_Summer_2018",
  save_csv = TRUE,
  n_rec = 0
)
}
\arguments{
\item{host}{The default is \code{https://gisp.dfo-mpo.gc.ca}.  This identifies the host  url for the service.}

\item{service}{The default value is
\code{'FGP/ADAPT_Canada_Atlantic_Summer_2018'}.  This identifies the folder and service from
which you want to extract data.}

\item{save_csv}{The default value is \code{TRUE}, which means that the
extracted data is saved to a csv file in your working directory.  If
\code{FALSE}, no csv will be created.}

\item{n_rec}{The default is \code{0}.  Primarily for debugging, this
allows the user to select only some records, rather than doing a full
extraction.}
}
\description{
This function facilitates the extraction of data from the DFO 
ESRI REST services into R objects and other formats.  
By default, the REST services only allow the extraction of 1000 records at a 
time, and this function does many successive extractions, and merges the
records together into a single object.
}
\seealso{
Other ArcGIS: 
\code{\link{getGeoData}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{ArcGIS}
