% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeoData.R
\name{getGeoData}
\alias{getGeoData}
\title{getGeoData}
\usage{
getGeoData(theService = NULL, filt = NULL, extr_Lim = 1000)
}
\arguments{
\item{theService}{This specifies which service is being requested. The 
default is \code{NULL}, but the only current, valid value is 
\code{"FGP/Groundfish_WFS_Service"}}

\item{filt}{the default is \code{NULL}. This specifies the table(s) you want 
data from, and the appropriate values for any particular field(s).  The format 
is \code{tableName1.fieldName1:<value1,value2>&tableName1.fieldName2:<value1,value2>&tableName3.fieldName1:<value1,value2>} 
Please see the example for a real example.}

\item{extr_Lim}{default is \code{1000}.  This identifies how many records can 
be extracted in a single pull from the WFS service.  Unless you know that the 
value is something else, it should be left at 1000.}
}
\description{
This function can pull data down from a GeoData service on the 
DFO Geoportal site.  Tables with filters must be specified, and these 
determine what data is returned.
}
\examples{
getGeoData(theService="FGP/Groundfish_WFS_Service", filt="Groundfish_GSSPECIES.CODE:10,11,12&Groundfish_GSMISSIONS.YEAR:2015,2016,2017&Groundfish_GSMISSIONS.SEASON:SUMMER")
}
\seealso{
Other ArcGIS: \code{\link{get_DFO_REST}}
}
\author{
Mike McMahon
}
\concept{ArcGIS}
