% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_DFO_REST.R
\name{get_DFO_REST}
\alias{get_DFO_REST}
\title{get_DFO_REST}
\usage{
get_DFO_REST(service = "ADAPT_Canada_Atlantic_Summer_2016", save_csv = TRUE,
  n_rec = 0)
}
\arguments{
\item{service}{The default value is
\code{'ADAPT_Canada_Atlantic_Summer_2016'}.  This identifies the service from
which you want to extract data.}

\item{save_csv}{The default value is \code{TRUE}, which means that the
extracted data is saved to a csv file in your working directory.  If
\code{FALSE}, no csv will be created.}

\item{n_rec}{The default is \code{0}.  Primarily for debugging, this
allows the user to select only some records, rather than doing a full
extraction.}
}
\description{
This function facilitates the extraction of data from the DFO 
ESRI REST services into R objects and other formats.  
By default, the REST services only allow the extraction of 1000 records at a 
time, and this function does many successive extractions, and merges the
records together into a single object.
}
\seealso{
Other ArcGIS: \code{\link{pullDFOGIS}},
  \code{\link{pullDFOREST}}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
